Combine charts components
