+++ src/app/app-routing.module.ts
@@ -6,6 +6,7 @@ import { ChartsComponent }   from './charts/charts.component';
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
  
 const appRoutes: Routes = [
+  { path: 'charts', component: ChartsComponent },
   { path: '**', component: PageNotFoundComponent }
 ];
