--- src/app/charts/bar-chart/bar-chart.service.ts
@@ -1,49 +1,0 @@
-import { Injectable } from '@angular/core';
-
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/of';
-
-const BAR_CHART_DATA = [
-  {
-    key: "sdadasd",
-    values: [
-      {
-        "label" : "A" ,
-        "value" : -29.765957771107
-      },
-      {
-        "label" : "B" ,
-        "value" : 0
-      },
-      {
-        "label" : "C" ,
-        "value" : 32.807804682612
-      },
-      {
-        "label" : "D" ,
-        "value" : 196.45946739256
-      },
-      {
-        "label" : "E" ,
-        "value" : 0.19434030906893
-      },
-      {
-        "label" : "F" ,
-        "value" : -98.079782601442
-      },
-      {
-        "label" : "G" ,
-        "value" : -13.925743130903
-      },
-      {
-        "label" : "H" ,
-        "value" : -5.1387322875705
-      }
-    ]
-  }
-];
-
-@Injectable()
-export class BarChartService {
-  getData() { return Observable.of(BAR_CHART_DATA); }
-}
