+++ src/app/charts/charts.component.html
@@ -1,6 +1,9 @@
-<h1>line chart</h1>
-<line-chart></line-chart>
-<h1>bar chart</h1>
-<bar-chart></bar-chart>
-<h1>pie chart</h1>
-<pie-chart></pie-chart>
+<div>
+  <nvd3 [options]="lineChart" [data]="lineChartData"></nvd3>
+</div>
+<div>
+  <nvd3 [options]="barChart" [data]="barChartData"></nvd3>
+</div>
+<div>
+  <nvd3 [options]="pieChart" [data]="pieChartData"></nvd3>
+</div>
