--- src/app/charts/pie-chart/pie-chart.component.html
@@ -1,3 +1,0 @@
-<div>
-  <nvd3 [options]="options" [data]="data"></nvd3>
-</div>
